+++ src/Armada.Core/Database/Sqlite/SqliteDatabaseDriver.cs
@@ -922,7 +922,6 @@ namespace Armada.Core.Database.Sqlite
             ProjectProfiles = new ProjectProfileMethods(this, _Settings, _Logging);
             Skills = new SkillMethods(this, _Settings, _Logging);
             CoordinationLeases = new CoordinationLeaseMethods(this, _Settings, _Logging);
-            Jobs = new JobMethods(this, _Settings, _Logging);
             TokenUsage = new TokenUsageMethods(this, _Settings, _Logging);
             ModelEndpoints = new ModelEndpointMethods(this, _Settings, _Logging);
         }